remove 4 web-components, in favor of 1 new <value-connected>
