+++ frontend/src/util.mjs
@@ -186,3 +186,37 @@ Object.byString = function(o, s) {
   }
   return o;
 }
+
+
+// stupid rainbow effect, replacing text
+//
+export const RAINBOWS = () => {
+  (function () {
+
+      var elems = document.getElementsByTagName('rainbow')
+
+      Array.prototype.forEach.call(elems, rain)
+
+      function rain(elem){
+        var angle = 0;
+        var text = elem.textContent.split('');
+        var len = text.length;
+        var phaseJump = 360 / len;
+        var spans;
+      
+        elem.innerHTML = text.map(function (char) {
+          return '<span class="text-rainbow">' + char + '</span>';
+        }).join('');
+      
+        spans = elem.children;
+      
+        (function wheee () {
+          for (var i = 0; i < len; i++) {
+            spans[i].style.color = 'hsl(' + (angle + Math.floor(i * phaseJump)) + ', 55%, 70%)';
+          }
+          angle++;
+          requestAnimationFrame(wheee);
+        })();
+      }
+    })();
+}
